AI Fixes and More
--- Player ---
• Fixed a bug that prevented the AI from making the best possible move.

• Fixed code for when playing on 'Infinity' mode.

--- CX File ---
• Small fixes to the ReadKey() method.

--- Gameboard ---
• Type conversion changes.

--- GameEngine ---
• Typo fixes.

--- Program ---
• Added Linq namespace.

+++ Connect4/CX.cs
@@ -28,9 +28,9 @@ namespace ConnectX
                 switch (playerChoice)
                 {
                     case ConsoleKey.Escape:
-                        return !menu ? 0 : 1;
+                        return !menu ? (byte)(0) : (byte)(1);
                     case ConsoleKey.P:
-                        return !menu ? 0 : 1;
+                        return !menu ? (byte)(0) : (byte)(1);
                     case ConsoleKey.D0:
                         return 0;
                     case ConsoleKey.D1:
